requirements.txt
numpy
pandas
tqdm
tensorflow
pickle
matplotlib
pillow